+++ src/databricks/labs/ucx/recon/data_profiler.py
@@ -12,6 +12,17 @@ class StandardDataProfiler(DataProfiler):
         self._metadata_retriever = metadata_retriever
 
     def profile_data(self, entity: TableIdentifier) -> DataProfilingResult:
+        """
+        This method profiles the data in the given table. It takes a TableIdentifier object as input, which represents
+        the table to be profiled. The method performs two main operations:
+
+        1. It retrieves the row count of the table.
+        2. It retrieves the metadata of the table using a TableMetadataRetriever instance.
+
+        Note: This method does not handle exceptions raised during the execution of the SQL query or the retrieval
+        of the table metadata. These exceptions are expected to be handled by the caller
+        in a manner appropriate for their context.
+        """
         row_count = self._get_table_row_count(entity)
         return DataProfilingResult(
             row_count,
@@ -19,10 +30,6 @@ class StandardDataProfiler(DataProfiler):
         )
 
     def _get_table_row_count(self, entity: TableIdentifier) -> int:
-        query_result: Iterator[Row] = self._sql_backend.fetch(
-            f"SELECT COUNT(*) as row_count FROM {entity.table}",
-            catalog=entity.catalog,
-            schema=entity.schema,
-        )
+        query_result: Iterator[Row] = self._sql_backend.fetch(f"SELECT COUNT(*) as row_count FROM {entity.fqn_escaped}")
         count_row = next(query_result)
         return int(count_row[0])
