+++ tests/unit/recon/test_data_comparator.py
@@ -1,6 +1,14 @@
+import re
+
 from databricks.labs.lsql.backends import MockBackend
 
-from databricks.labs.ucx.recon.base import TableIdentifier, DataComparisonResult
+from databricks.labs.ucx.recon.base import (
+    TableIdentifier,
+    DataComparisonResult,
+    DataProfilingResult,
+    TableMetadata,
+    ColumnMetadata,
+)
 from databricks.labs.ucx.recon.data_comparator import StandardDataComparator
 from databricks.labs.ucx.recon.data_profiler import StandardDataProfiler
 from databricks.labs.ucx.recon.metadata_retriever import DatabricksTableMetadataRetriever
@@ -19,8 +27,8 @@ def test_data_comparison(metadata_row_factory, row_count_row_factory, data_comp_
                 ("col1", "int"),
                 ("col2", "string"),
             ],
-            f"SELECT COUNT\\(\\*\\) as row_count FROM {source.table}": row_count_row_factory[100,],
-            f"SELECT COUNT\\(\\*\\) as row_count FROM {target.table}": row_count_row_factory[2,],
+            f"SELECT COUNT\\(\\*\\) as row_count FROM {source.fqn_escaped}": row_count_row_factory[100,],
+            f"SELECT COUNT\\(\\*\\) as row_count FROM {target.fqn_escaped}": row_count_row_factory[2,],
             "WITH compare_results": data_comp_row_factory[(102, 100, 2),],
         }
     )
@@ -37,3 +45,64 @@ def test_data_comparison(metadata_row_factory, row_count_row_factory, data_comp_
     actual_comparison_result = data_comparator.compare_data(source, target)
 
     assert actual_comparison_result == expected_comparison_result
+
+
+def test_prepare_data_comparison_query():
+    source = TableIdentifier("hive_metastore", "db1", "table1")
+    target = TableIdentifier("catalog1", "schema1", "table2")
+
+    source_data_profile = DataProfilingResult(
+        10,
+        TableMetadata(
+            source,
+            [
+                ColumnMetadata("col1", "string"),
+                ColumnMetadata("col2", "array<string>"),
+                ColumnMetadata("col3", "struct<a:int,b:int,c:array<string>>"),
+            ],
+        ),
+    )
+    target_data_profile = DataProfilingResult(
+        10,
+        TableMetadata(
+            target,
+            [
+                ColumnMetadata("col1", "string"),
+                ColumnMetadata("col2", "array<string>"),
+                ColumnMetadata("col3", "struct<a:int,b:int,c:array<string>>"),
+            ],
+        ),
+    )
+
+    actual_query = (
+        StandardDataComparator.build_data_comparison_query(
+            source_data_profile,
+            target_data_profile,
+        )
+        .strip()
+        .lower()
+    )
+
+    source_hash_columns = [
+        "COALESCE(TRIM(col1), '')",
+        "COALESCE(TRIM(TO_JSON(SORT_ARRAY(col2))), '')",
+        "COALESCE(TRIM(TO_JSON(col3)), '')",
+    ]
+    target_hash_columns = [
+        "COALESCE(TRIM(col1), '')",
+        "COALESCE(TRIM(TO_JSON(SORT_ARRAY(col2))), '')",
+        "COALESCE(TRIM(TO_JSON(col3)), '')",
+    ]
+
+    expected_query = (
+        StandardDataComparator.DATA_COMPARISON_QUERY_TEMPLATE.format(
+            source_hash_expr=f"SHA2(CONCAT_WS('|', {', '.join(source_hash_columns)}), 256)",
+            target_hash_expr=f"SHA2(CONCAT_WS('|', {', '.join(target_hash_columns)}), 256)",
+            source_table_fqn="`hive_metastore`.`db1`.`table1`",
+            target_table_fqn="`catalog1`.`schema1`.`table2`",
+        )
+        .strip()
+        .lower()
+    )
+
+    assert re.sub(r'\s+', ' ', actual_query) == re.sub(r'\s+', ' ', expected_query)
