+++ tests/unit/recon/test_metadata_retriever.py
@@ -1,3 +1,5 @@
+import re
+
 from databricks.labs.lsql.backends import MockBackend
 
 from databricks.labs.ucx.recon.base import TableIdentifier, TableMetadata, ColumnMetadata
@@ -37,7 +39,7 @@ def test_unity_table_metadata_retrieval(metadata_row_factory):
     table_identifier = TableIdentifier("catalog1", "db1", "table1")
     sql_backend = MockBackend(
         rows={
-            f"{table_identifier.catalog}.information_schema.columns": metadata_row_factory[
+            f"{table_identifier.catalog_escaped}.information_schema.columns": metadata_row_factory[
                 ("col2", "string"),
                 ("col1", "int"),
                 ("col3", "array<string>"),
@@ -57,3 +59,29 @@ def test_unity_table_metadata_retrieval(metadata_row_factory):
     metadata_retriever = DatabricksTableMetadataRetriever(sql_backend)
     actual_metadata = metadata_retriever.get_metadata(table_identifier)
     assert actual_metadata == expected_metadata
+
+
+def test_hms_metadata_query():
+    table_identifier = TableIdentifier("hive_metastore", "db1", "table1")
+    actual_query = DatabricksTableMetadataRetriever.build_metadata_query(table_identifier).strip().lower()
+    expected_query = "DESCRIBE TABLE `hive_metastore`.`db1`.`table1`".lower()
+    assert re.sub(r'\s+', ' ', actual_query) == expected_query
+
+
+def test_unity_metadata_query():
+    table_identifier = TableIdentifier("catalog1", "db1", "table1")
+    actual_query = DatabricksTableMetadataRetriever.build_metadata_query(table_identifier).strip().lower()
+    expected_query = """
+        SELECT 
+            LOWER(column_name) AS col_name, 
+            full_data_type AS data_type
+        FROM 
+            `catalog1`.information_schema.columns
+        WHERE
+            LOWER(table_catalog)='catalog1' AND
+            LOWER(table_schema)='db1' AND
+            LOWER(table_name) ='table1'
+        ORDER BY col_name
+    """.strip().lower()
+
+    assert re.sub(r'\s+', ' ', actual_query) == re.sub(r'\s+', ' ', expected_query)
